Add new volume request to calculate the volume for a given shape
User will have to pass a few arguments depending what shape is used.
Volume can be calculated for only a cube or a cylinder. A cube will only
need a edge value where as a cylinder will need radius and height. This
addition will figure out which calculation to perform depending on the shape.
Once the shape is calculated the results are diplayed to the user. The Java
Math library was used to perform calculations.